Boardコンポーネントを一元管理化
