+++ types/index.ts
@@ -5,5 +5,6 @@ export type Coordinates = { x: number; y: number } | null;
 export type BoardProps = {
   xIsNext: boolean;
   squares: Mark[];
+  is3x3: boolean;
   onPlay: (nextSquares: Mark[], coordinates: Coordinates) => void;
 };
